+++ src/nlcbsmm/page.h
@@ -13,12 +13,12 @@ namespace NLCBSMM {
 
    class Page {
       public:
-         uint32_t address;
-         uint32_t protection;
-         uint32_t version;
-         uint32_t dirty;
+         intptr_t address;
+         int protection;
+         int version;
+         int dirty;
 
-         Page(uint32_t _address, uint32_t _protection) {
+         Page(intptr_t _address, int _protection) {
             address    = _address;
             protection = _protection;
             version    = 0;
